# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alpha Bit Gadget Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/offers:
    get:
      operationId: getOfferProducts
      tags: [products]
      summary: Get products on sale
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Offer products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ── CATEGORIES ────────────────────────────────────────────────────────────
...
  /payments:
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment link via AlphaBitPay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Payment link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
        "400":
          description: Bad request
  /payments/status/{orderId}:
    get:
      operationId: getPaymentStatus
      tags: [payments]
      summary: Get payment status for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentStatus"

  # ── AUTH ──────────────────────────────────────────────────────────────────
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List products (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductList"
    post:
      operationId: createAdminProduct
      tags: [admin]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{productId}:
    put:
      operationId: updateAdminProduct
      tags: [admin]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteAdminProduct
      tags: [admin]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required: [id, name, description, short_description, price, category, image, stock, sku]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: string
        stock:
          type: integer
        sku:
          type: string
        featured:
          type: boolean
        on_sale:
          type: boolean
    ProductList:
      type: object
      required: [products, total, limit, skip]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        skip:
          type: integer
    ProductInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        stock:
          type: integer
          default: 100
        sku:
          type: string
        featured:
          type: boolean
        on_sale:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        on_sale:
          type: boolean
    AdminProductList:
      type: object
      required: [products, categories, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        categories:
          type: array
          items:
            type: string
        total:
          type: integer

    # ── Category ──────────────────────────────────────────────────────────────
...
    PaymentInput:
      type: object
      required: [order_id]
      properties:
        order_id:
          type: string
        payment_method:
          type: string
          default: paypal
    PaymentResult:
      type: object
      required: [checkout_url, payment_id]
      properties:
        checkout_url:
          type: string
        payment_id:
          type: string
    PaymentStatus:
      type: object
      required: [order_id, status]
      properties:
        order_id:
          type: string
        status:
          type: string
        payment_id:
          type: ["string", "null"]

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string